replay: Let replay_char_read_all_load() use size_t

+++ include/sysemu/replay.h
@@ -160,7 +160,7 @@ void replay_char_write_event_save(int res, int offset);
 /*! Reads char write return value from the replay log. */
 void replay_char_write_event_load(int *res, size_t *offset);
 /*! Reads information about read_all character event. */
-int replay_char_read_all_load(uint8_t *buf);
+size_t replay_char_read_all_load(uint8_t *buf);
 /*! Writes character read_all error code into the replay log. */
 void replay_char_read_all_save_error(int res);
 /*! Writes character read_all execution result into the replay log. */
